+++ CHANGELOG.md
@@ -11,6 +11,7 @@ SOUP version numbering generally follows the _major.minor.patch_ scheme.  Odd mi
 * [mse242944: Long display name with no spaces breaks design of review history pages](http://meta.stackexchange.com/q/242944) (CSS)
 * [mse244587: “Top Network Users” should contain themselves!](http://meta.stackexchange.com/q/244587) (CSS, stackexchange.com base site only)
 * [rpg3554: Skin does not handle read-only mode well](http://meta.rpg.stackexchange.com/q/3554) (CSS, RPG.SE only)
+* [mse240787: Inconsistent reputation mouse-over text](http://meta.stackexchange.com/q/240787)
 
 1.32 (13 Aug 2015)
 ====